# docs/swagger.yaml
definitions:
...
  handlers.WebSocketMessage:
    properties:
      payload: {}
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /ws:
    get:
      description: Upgrades HTTP connection to WebSocket protocol for real-time communication
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "426":
          description: Upgrade Required - Client must request WebSocket upgrade
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: WebSocket connection upgrade
      tags:
      - websocket
  /ws/broadcast:
    post:
      consumes:
      - application/json
      description: Sends a real-time notification to all connected users via WebSocket
      parameters:
      - description: Type of notification
        in: body
        name: notification_type
        required: true
        schema:
          type: string
      - description: Notification payload
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: Notification broadcasted successfully
          schema:
            $ref: '#/definitions/handlers.WebSocketMessage'
      security:
      - BearerAuth: []
      summary: Broadcast notification to all users
      tags:
      - websocket
  /ws/notify/{user_id}:
    post:
      consumes:
      - application/json
      description: Sends a real-time notification to a specific user via WebSocket
      parameters:
      - description: User ID to send notification to
        in: path
        name: user_id
        required: true
        type: integer
      - description: Type of notification
        in: body
        name: notification_type
        required: true
        schema:
          type: string
      - description: Notification payload
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: Notification sent successfully
          schema:
            $ref: '#/definitions/handlers.WebSocketMessage'
        "404":
          description: User not connected
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send notification to user
      tags:
      - websocket
swagger: "2.0"